research_platform/requirements.txt
# Web framework & REST
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-extensions>=3.2
django-filter>=23.0

# Database
mysqlclient>=2.2

# Async / task queue
celery>=5.3
redis>=5.0
channels>=4.0
channels-redis>=4.1

# Search
elasticsearch-dsl>=8.0

# ML / NLP
transformers>=4.38
torch>=2.2
sentence-transformers>=2.5
scikit-learn>=1.4
peft>=0.9
tensorflow>=2.15
numpy>=1.26
pandas>=2.2

# PDF processing
PyPDF2>=3.0
pypdf>=3.0

# Image processing
Pillow>=10.0

# Visualisation (dev / reporting only)
matplotlib>=3.8
seaborn>=0.13

# Utilities
networkx>=3.2

# RAG / LangGraph
langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
langchain-community>=0.3
langgraph>=0.2
chromadb>=0.5
python-dotenv>=1.0
